refatore(modal): moved note modal logic to another file
