+++ src/scripts/navbar/types/htmlElements.ts
@@ -1,3 +1,4 @@
 export type DivElement = HTMLDivElement;
 export type BtnElement = HTMLButtonElement;
 export type InputElement = HTMLInputElement;
+export type UListElement = HTMLUListElement;
